move clearday to own component
