+++ src/views/NewPlan/PlannedDinner.tsx
@@ -8,6 +8,7 @@ import {
   DialogTitle,
   DialogDescription,
 } from "../../components/ui/dialog";
+import { ClearDay } from "./ClearDay";
 
 type Props = {
   dinner: DinnerWithTags;
@@ -61,57 +62,5 @@ export const PlannedDinner = ({ dinner, date, closeDialog }: Props) => {
         </div>
       </div>
     </>
-  );
-};
-
-type ClearDayProps = { date: Date; closeDialog: () => void };
-
-const ClearDay = ({ date, closeDialog }: ClearDayProps) => {
-  const posthog = usePostHog();
-  const utils = api.useUtils();
-  const unplanDayMutation = api.plan.unplanDay.useMutation({
-    onMutate: (input) => {
-      void utils.plan.plannedDinners.cancel();
-      const prevPlannedDinners = utils.plan.plannedDinners.getData();
-      utils.plan.plannedDinners.setData(
-        { startOfWeek: startOfWeek(date ?? new Date(), { weekStartsOn: 1 }) },
-        (old) => {
-          return {
-            plans:
-              old?.plans.filter((plan) => !isSameDay(plan.date, input.date)) ??
-              [],
-          };
-        },
-      );
-      return { prevPlannedDinners };
-    },
-    onError: (_, __, context) => {
-      if (context?.prevPlannedDinners) {
-        utils.plan.plannedDinners.setData(
-          { startOfWeek: startOfWeek(date ?? new Date(), { weekStartsOn: 1 }) },
-          context.prevPlannedDinners,
-        );
-      }
-    },
-    onSettled: async () => {
-      await utils.plan.plannedDinners.invalidate();
-    },
-    onSuccess: (res) => {
-      posthog.capture("clear day", { day: format(res.deleted.date, "EEE do") });
-      closeDialog();
-    },
-  });
-  return (
-    <Button
-      variant={"outline"}
-      onClick={() =>
-        unplanDayMutation.mutate({
-          date,
-          secret: localStorage.getItem("sulten-secret"),
-        })
-      }
-    >
-      Clear day
-    </Button>
   );
 };
